game\shared\portal: Use x86-64 indexer for GetEntityPhysicsObjects

+++ game/shared/portal/PortalSimulation.cpp
@@ -99,7 +99,7 @@ static int s_iPortalSimulatorGUID = 0; //used in standalone function that have n
 static void ConvertBrushListToClippedPolyhedronList( const int *pBrushes, int iBrushCount, const float *pOutwardFacingClipPlanes, int iClipPlaneCount, float fClipEpsilon, CUtlVector<CPolyhedron *> *pPolyhedronList );
 static void ClipPolyhedrons( CPolyhedron * const *pExistingPolyhedrons, int iPolyhedronCount, const float *pOutwardFacingClipPlanes, int iClipPlaneCount, float fClipEpsilon, CUtlVector<CPolyhedron *> *pPolyhedronList );
 static inline CPolyhedron *TransformAndClipSinglePolyhedron( CPolyhedron *pExistingPolyhedron, const VMatrix &Transform, const float *pOutwardFacingClipPlanes, int iClipPlaneCount, float fCutEpsilon, bool bUseTempMemory );
-static int GetEntityPhysicsObjects( IPhysicsEnvironment *pEnvironment, CBaseEntity *pEntity, IPhysicsObject **pRetList, int iRetListArraySize );
+static intp GetEntityPhysicsObjects( IPhysicsEnvironment *pEnvironment, CBaseEntity *pEntity, IPhysicsObject **pRetList, int iRetListArraySize );
 static CPhysCollide *ConvertPolyhedronsToCollideable( CPolyhedron **pPolyhedrons, int iPolyhedronCount );
 
 #ifndef CLIENT_DLL
@@ -2692,9 +2692,9 @@ static inline CPolyhedron *TransformAndClipSinglePolyhedron( CPolyhedron *pExist
 	return pNewPolyhedron;
 }
 
-static int GetEntityPhysicsObjects( IPhysicsEnvironment *pEnvironment, CBaseEntity *pEntity, IPhysicsObject **pRetList, int iRetListArraySize )
+static intp GetEntityPhysicsObjects( IPhysicsEnvironment *pEnvironment, CBaseEntity *pEntity, IPhysicsObject **pRetList, int iRetListArraySize )
 {	
-	int iCount, iRetCount = 0;
+	intp iCount, iRetCount = 0;
 	const IPhysicsObject **pList = pEnvironment->GetObjectList( &iCount );
 
 	if( iCount > iRetListArraySize )
